requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
Request
jsonify
from passlib.hash import pbkdf2_sha256